assignment done
computes the inverse of a matrix but in a weird way.  if you call
cacheSolve, you must pass it the vector created by makeCacheMatrix… the
best way is to set some object to the output of makeCacheMatrix, then
run cacheSolve on the object.  If you try to put makeCacheMatrix into
the function argument of cacheSolve, you aren’t doing anything more
efficiently